Refactored each react component to be in its own file
