+++ frontend/src/DeviceList.js
@@ -23,206 +23,16 @@ import {
 } from '@material-ui/icons/';
 
 import PropTypes from "prop-types";
-import GenerateData from "./SampleData";
-import HeadCells from "./HeadCells";
-import ActionMenu from "./ActionMenu";
-import DeviceListTabs from "./DeviceListTabs";
+import GenerateData from "./devicelist/SampleData";
+import HeadCells from "./devicelist/HeadCells";
+import DeviceListTabs from "./devicelist/VerticalTabs";
 import * as DeviceApi from "./api/DeviceApi";
-import SearchBar from './DeviceListSearch';
-import SortBy from './DeviceListSortBy';
-
-// imports for material ui & etc
-
-// temporary row
-
-function getStatusStyle(status) {
-    switch(status){
-        case 0:
-            return "green statusText";
-        case 1:
-            return "yellow statusText";
-        case 2:
-            return "red statusText";
-        default:
-            return "lightGrey statusOfflineText";   
-    }
-}
-function getStatusText(status) {
-    switch (status) {
-        case 0:
-            return "Online";
-        case 1:
-            return "Pending";
-        case 2:
-            return "Error";
-        default:
-            return "Offline";
-    }
-}
-
-class TitleBox extends React.Component {
-    render() {
-        return (
-            <React.Fragment>
-                <Box class="flexContents headerArea">
-                    <div class="paddedText title">My Devices</div>
-                    <div class="alignRightFloat">
-                        <Button class="green buttonText">
-                            <SwapHoriz /> Stream
-                        </Button>
-                        <Button class="blue buttonText">
-                            <AddSharp /> Add Device
-                        </Button>
-                    </div>
-                </Box>
-            </React.Fragment>
-        );
-    }
-}
-
-// tabs. decide on sender or receiver table
-class ContentsTable extends React.Component {
-    constructor(props) {
-        super(props)
-        var sampleData = GenerateData();
-        this.state = {
-            senders: sampleData.senders,
-            receivers: sampleData.receivers,
-            value: 0
-        }
-        this.handleValueChange = this.handleValueChange.bind(this);
-        this.handleSendersChange = this.handleSendersChange.bind(this);
-        this.handleReceiversChange = this.handleReceiversChange.bind(this);
-    }
-
-    componentDidMount() {
-        DeviceApi.getSenders(this.handleSendersChange);
-        DeviceApi.getReceivers(this.handleReceiversChange);
-    }
-
-    handleValueChange(value) {
-        this.setState({
-            value: value
-        });
-    }
-
-    handleSendersChange(senders) {
-        this.setState({
-            senders: senders
-        });
-    }
-
-    handleReceiversChange(receivers) {
-        this.setState({
-            receivers: receivers
-        });
-    }
-
-    render() {
-        return (
-            <React.Fragment>
-                <Box style={{display: 'flex', flexGrow: 1, margin: "1em 0em"}}>
-                    <SearchBar/>
-                    <SortBy/>
-                </Box>
-                <Box style={{display: 'flex', flexGrow: 1, maxHeight: 500}}>
-                    {DeviceListTabs(this.props.classes, [this.state.value, this.handleValueChange])}
-                    <TabPanel value={this.state.value} index={0}>
-                        <DevicesTable rows={this.state.senders} />
-                    </TabPanel>
-                    <TabPanel value={this.state.value} index={1}>
-                        <DevicesTable rows={this.state.receivers} />
-                    </TabPanel>
-                </Box>
-            </React.Fragment>
-        );
-    }
-}
-
-class SingleTableRow extends React.Component {
-    constructor(props) {
-        super(props)
-        this.state = {
-            open: false
-        }
-    }
-
-    rowExtras(extras) {
-        var extraStr = "";
-        for (var i = 0; i < extras.length; i++) {
-            extraStr = extraStr + extras[i] + " ";
-        }
-        return extraStr;
-    }
-
-    render() {
-        return (
-            <React.Fragment>
-                <TableRow key={this.props.row.id}>
-                    <TableCell class = "tableCell" style={{ width: 1, padding: 0, paddingLeft: 5 }}>
-                        <IconButton onClick={() => this.setState({ open: !this.state.open })}>
-                            {this.state.open ? <ExpandLess />:<ExpandMore />}
-                        </IconButton>
-                    </TableCell>
-                    <TableCell class = "tableCell">{this.props.row.name}</TableCell>
-                    <TableCell class = "tableCell numeric">{this.props.row.serial}</TableCell>
-                    <TableCell class = "tableCell">
-                        <div class={getStatusStyle(this.props.row.status)}>
-                            {getStatusText(this.props.row.status)}
-                        </div>
-                    </TableCell>
-                    <TableCell class = "tableCell numeric">{this.props.row.ip}</TableCell>
-                    <TableCell class = "tableCell numeric">{this.props.row.port}</TableCell>
-                    <TableCell class = "tableCell" align="center">
-                        <ActionMenu />
-                    </TableCell>
-                </TableRow>
-                <TableRow>
-                    <TableCell class="chevronText lightestGrey" colSpan={7}>
-                        <Collapse in={this.state.open} timeout="auto" unmountOnExit>
-                            <Box margin={2}>
-                                <Typography variant="caption">
-                                    {this.rowExtras(this.props.row.extras)}
-                                </Typography>
-                            </Box>
-                        </Collapse>
-                    </TableCell>
-                </TableRow>
-            </React.Fragment>
-        );
-    }
-}
-
-class DevicesTable extends React.Component {
-    constructor(props) {
-        super(props)
-    }
-
-    render() {
-        return (
-            <React.Fragment>
-                <Box>
-                    <TableContainer style={{ maxHeight: 500 }}>
-                        <Table stickyHeader>
-                            <TableHead>
-                                <TableRow>
-                                    <TableCell class="tableHeadCell lightestGrey" style={{ width: 1, padding: 0, paddingLeft: 5 }}></TableCell>
-                                    <HeadCells/>
-                                    <TableCell class="tableHeadCell lightestGrey" align="center"><Typography variant="caption">Actions</Typography></TableCell>
-                                </TableRow>
-                            </TableHead>
-                            <TableBody>
-                                {this.props.rows.map((row) => {
-                                    return <SingleTableRow row={row} />;
-                                })}
-                            </TableBody>
-                        </Table>
-                    </TableContainer>
-                </Box>
-            </React.Fragment>
-        );
-    }
-}
+import SearchBar from './devicelist/SearchBar';
+import SortBy from './devicelist/SortBy';
+import SingleTableRow from './devicelist/SingleTableRow'
+import TitleBox from './devicelist/TitleBox'
+import DevicesTable from './devicelist/DevicesTable'
+import ContentsTable from './devicelist/ContentsTable'
 
 // combine the fragments
 class DeviceList extends React.Component {
@@ -239,32 +49,6 @@ class DeviceList extends React.Component {
         );
     }
 }
-
-function TabPanel(props) {
-    const { children, value, index, ...other } = props;
-
-    return (
-        <div
-            role="tabpanel"
-            hidden={value !== index}
-            id={`vertical-tabpanel-${index}`}
-            aria-labelledby={`vertical-tab-${index}`}
-            {...other}
-        >
-            {value === index && (
-                <Box p={0}>
-                    <Typography component={'span'}>{children}</Typography>
-                </Box>
-            )}
-        </div>
-    );
-}
-
-TabPanel.propTypes = {
-    children: PropTypes.node,
-    index: PropTypes.any.isRequired,
-    value: PropTypes.any.isRequired
-};
 
 const useStyles = (theme) => ({
     root: {
