+++ frontend/src/api/DeviceApi.js
@@ -6,7 +6,7 @@ export function getSenders(callback) {
         .then((senders) => {
             callback(
                 senders.data.map((sender) => {
-                    return new DeviceInfo("id", sender.device.displayName, sender.serialNumber, "Online", "127.0.0.1", "2300", "Sample sender");
+                    return new DeviceInfo("id", sender.device.displayName, sender.serialNumber, 0, "127.0.0.1", "2300", "Sample sender");
             }));
         })
         .catch((error) => {
@@ -18,7 +18,7 @@ export function getReceivers(callback) {
     axios("http://localhost:8080/decoder")
         .then((receivers) => {
             callback(receivers.data.map((sender) => {
-                return new DeviceInfo("id", sender.device.displayName, sender.serialNumber, "Online", "127.0.0.1", "2400", "Sample receiver");
+                return new DeviceInfo("id", sender.device.displayName, sender.serialNumber, 0, "127.0.0.1", "2400", "Sample receiver");
             }));
         })
         .catch((error) => {
