+++ frontend/src/devicelist/HeadCells.js
