+++ frontend/src/devicelist/SearchBar.js
