+++ frontend/src/devicelist/TabPanel.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import {
+    Box,
+    Button,
+    Collapse,
+    Container,
+    Table,
+    TableHead,
+    TableRow,
+    TableCell,
+    TableContainer,
+    TableBody,
+    Typography,
+    IconButton,
+    withStyles
+} from "@material-ui/core"
+
+export default function TabPanel(props) {
+    const { children, value, index, ...other } = props;
+
+    return (
+        <div
+            role="tabpanel"
+            hidden={value !== index}
+            id={`vertical-tabpanel-${index}`}
+            aria-labelledby={`vertical-tab-${index}`}
+            {...other}
+        >
+            {value === index && (
+                <Box p={0}>
+                    <Typography component={'span'}>{children}</Typography>
+                </Box>
+            )}
+        </div>
+    );
+}
