+++ frontend/src/devicelist/VerticalTabs.js
